academycity/requirements.txt
lxml==4.6.1
#numpy==1.19.2
#pandas==1.1.3
scipy==1.5.3
numpy==1.19.1
pandas==1.0.5
# Pillow==7.2.0
psycopg2-binary==2.8.6
aioredis==1.3.1
amqp==2.5.1
appdirs==1.4.3
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
Babel==2.7.0
backport-collections==0.1
billiard==3.6.1.0
boto3==1.16.4
botocore==1.19.4
braintree==3.56.0
cairocffi==1.1.0
CairoSVG==2.4.2
celery==4.3.0
certifi==2019.6.16
cffi==1.13.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
constantly==15.1.0
cryptography==3.1.1
cssselect2==0.2.2
daphne==2.5.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.10
django-appconf==1.0.3
django-appdata==0.3.1
django-braces==1.14.0
django-ckeditor==5.7.1
django-ckeditor-widget==0.1
django-classy-tags==2.0.0
django-cms==3.7.4
django-currentuser==0.5.1
django-dbbackup==3.3.0
django-debug-toolbar==2.0
django-filer==2.0.2
django-formtools==2.2
django-js-asset==1.2.2
django-mptt==0.11.0
django-parler==2.2
django-phonenumber-field==1.3.0
django-polymorphic==3.0.0
django-reversion==3.0.4
django-rosetta==0.9.3
django-sekizai==2.0.0
django-sortedm2m==1.5.0
django-storages==1.7.1
django-taggit==1.1.0
django-treebeard==4.3.1
djangocms-admin-style==1.5.0
djangocms-attributes-field==2.0.0
djangocms-bootstrap4==1.6.0
djangocms-column==1.11.0
djangocms-file==2.4.0
djangocms-googlemap==1.4.0
djangocms-icon==1.5.0
djangocms-installer==1.2.3
djangocms-link==2.6.1
djangocms-picture==2.4.0
djangocms-snippet==2.3.0
djangocms-style==2.3.0
djangocms-text-ckeditor==3.10.0
djangocms-video==2.3.0
djangorestframework==3.9.4
dropbox==9.4.0
easy-thumbnails==2.7
et-xmlfile==1.0.1
eventkit==0.8.5
flower==0.9.3
gunicorn==20.0.4
hiredis==1.1.0
html5lib==1.1
hyperlink==20.0.1
ib-insync==0.9.47
idna==2.8
importlib-metadata==0.21
incremental==17.5.0
jdcal==1.4.1
jmespath==0.10.0
kombu==4.6.4
Markdown==2.6.11
more-itertools==8.5.0
msgpack==0.6.2

nest-asyncio==1.4.1
oauthlib==3.0.2
openpyxl==3.0.1
packaging==19.0
phonenumbers==8.10.15
phonenumberslite==8.10.15
polib==1.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
pyparsing==2.4.0
Pyphen==0.9.5
python-dateutil==2.8.0
python-slugify==3.0.2
python3-openid==3.1.0
pytz==2020.1
redis==3.3.8
requests==2.24.0
requests-oauthlib==1.2.0
s3transfer==0.3.3
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.2
tinycss2==1.0.2
toml==0.10.0
tornado==5.1.1
Twisted==20.3.0
txaio==20.4.1
tzlocal==2.1
Unidecode==1.1.1
urllib3==1.25.11
vine==1.3.0
WeasyPrint==50
webencodings==0.5.1
xlrd==1.2.0
YURL==1.0.0
zipp==3.3.2
zope.interface==5.1.2
matplotlib==3.3.0
sklearn==0.0

django-mathfilters==1.0.0

django-cookies-samesite
#
elasticsearch==7.13.1
elasticsearch_dsl==7.3.0
django-elasticsearch-dsl==7.2.0
#
simpy==4.0.1
mesa==0.8.9
#
